requirements.txt
asgiref==3.8.1
Django==5.0.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
PyMySQL==1.1.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
